requirements.txt
altair==5.3.0
annotated-types==0.6.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
catalogue==2.0.10
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
docker-pycreds==0.4.0
executing==2.0.1
fastcore==1.5.33
fastjsonschema==2.19.1
filelock==3.14.0
fonttools==4.51.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
icecream==2.1.3
identify==2.5.36
idna==3.7
imageio==2.34.1
imgaug==0.4.0
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lazy_loader==0.4
Levenshtein==0.25.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
pre-commit==3.7.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.0
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
sentry-sdk==2.0.1
setproctitle==1.3.3
setuptools==69.5.1
shapely==2.0.4
six==1.16.0
smmap==5.0.1
soupsieve==2.5
srsly==2.4.8
stack-data==0.6.3
threadpoolctl==3.5.0
tifffile==2024.5.3
tinycss2==1.3.0
toolz==0.12.1
torch-snippets==0.528
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.1
wandb==0.16.6
wasabi==1.1.2
wcwidth==0.2.13
webencodings==0.5.1
xmltodict==0.13.0
